Distribute client methods to service classes.
Services will be accessed as cached properties from the client instance.

Reconfigure tests to handle the new structure.

Update example scripts to use the new API.
