+++ tests/conftest.py
@@ -1,0 +1,16 @@
+"""Configuration for pytest."""
+
+from unittest.mock import AsyncMock
+
+import pytest
+
+
+class MockSession:
+    def __init__(self) -> None:
+        self.headers = {}
+        self.get = AsyncMock()
+
+
+@pytest.fixture
+def session():
+    return MockSession()
